Added getUserRolesCount and fixed connection null error
Created getUserRolesCount() by using getUserRole() and modified it so it returns an int for the roles counted that are tied to the user instead of a string. Still working on it and may need to use a different method instead.

Also fixed an exception error issue where the connection was null, so I included code to fix that issue in getUserRole().