Updated DatabaseHelper.java to add invite code expiration functionality
This commit adds functionality to the invite code table so that it can store the creation date of code. The creation date is then compared to the current time of the user that claims the code, if the code was generated more than 15 minutes prior then the code is no longer valid.